# docs/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: Learning Loop Agent API
  version: 0.2.0
  description: |
    AI 学习工具的目标 HTTP 契约。它覆盖 Ticket 01-16，Ticket 17-21 复用其中定义的工作流契约，
    是后端实现和前端集成的共同输入，
    不表示当前后端已经完成。除文件上传与下载外，请求和响应均使用 application/json。
    v0.2 使用 chat_style 表示无阶段门禁的对话风格，并用 api_format 显式选择模型调用协议。
    旧 learning_mode、SocraticState 和阶段型 intent 只用于迁移期兼容。
  x-contract-status: target
servers:
  - url: http://127.0.0.1:4173
    description: 本地开发服务
security: []
...
paths:
...
  /api/pet:
    get:
      operationId: getPet
      tags: [Pet]
      summary: 读取桌宠状态
      description: 首次访问时写入 adopted_at。墨水与统计从各科目已有数据只读推导。
      x-tickets: ['34']
      responses:
        '200':
          description: 当前桌宠
          content:
            application/json:
              schema:
                $ref: ./schemas.yaml#/components/schemas/Pet
    patch:
      operationId: updatePet
      tags: [Pet]
      summary: 更新桌宠显示设置
      description: name 去首尾空白后须为 1–12 个字符；position_x / position_y 夹取到 0–100。
      x-tickets: ['34']
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./schemas.yaml#/components/schemas/PetPatch
      responses:
        '200':
          description: 更新后的桌宠
          content:
            application/json:
              schema:
                $ref: ./schemas.yaml#/components/schemas/Pet
        '422': {$ref: '#/components/responses/ValidationFailed'}
...
  /api/subjects/{subject_id}/sources:
    parameters:
      - $ref: '#/components/parameters/SubjectId'
    get:
      operationId: listSources
      tags: [Sources]
      summary: 列出资料库中的用户资料
      x-tickets: ['03', '05']
      responses:
        '200':
          description: 用户资料列表
          content:
            application/json:
              schema:
                $ref: ./schemas.yaml#/components/schemas/SourceList
        '404': {$ref: '#/components/responses/NotFound'}
    post:
      operationId: uploadSource
      tags: [Sources, Operations]
      summary: 上传用户资料并开始解析
      description: 支持 Markdown、TXT、PDF、DOCX、PPTX 和常见图片。相同内容复用解析缓存。
      x-tickets: ['03', '05']
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: ./schemas.yaml#/components/schemas/SourceUpload
      responses:
        '202': {$ref: '#/components/responses/OperationAccepted'}
        '404': {$ref: '#/components/responses/NotFound'}
        '413':
          description: 文件超过允许大小
          content:
            application/json:
              schema: {$ref: ./schemas.yaml#/components/schemas/ErrorResponse}
        '415':
          description: 不支持的文件类型
          content:
            application/json:
              schema: {$ref: ./schemas.yaml#/components/schemas/ErrorResponse}
        '422': {$ref: '#/components/responses/ValidationFailed'}
  /api/sources/{source_id}:
    parameters:
      - $ref: '#/components/parameters/SourceId'
    get:
      operationId: getSource
      tags: [Sources]
      summary: 读取用户资料
      x-tickets: ['03', '05']
      responses:
        '200':
          description: 用户资料
          content:
            application/json:
              schema: {$ref: ./schemas.yaml#/components/schemas/Source}
        '404': {$ref: '#/components/responses/NotFound'}
    delete:
      operationId: deleteSource
      tags: [Sources]
      summary: 从资料库删除用户资料
      description: 停止后续检索；已被会话、试卷或学习产物引用的资料版本和来源引用仍可识别。
      x-tickets: ['03', '04']
      responses:
        '204': {description: 已从资料库删除}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/sources/{source_id}/versions:
    parameters:
      - $ref: '#/components/parameters/SourceId'
    get:
      operationId: listSourceVersions
      tags: [Sources]
      summary: 列出资料版本
      x-tickets: ['03', '05']
      responses:
        '200':
          description: 资料版本列表
          content:
            application/json:
              schema: {$ref: ./schemas.yaml#/components/schemas/SourceVersionList}
        '404': {$ref: '#/components/responses/NotFound'}
    post:
      operationId: createSourceVersion
      tags: [Sources, Operations]
      summary: 上传变更内容并创建资料版本
      x-tickets: ['03', '05']
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: {$ref: ./schemas.yaml#/components/schemas/SourceVersionUpload}
      responses:
        '202': {$ref: '#/components/responses/OperationAccepted'}
        '404': {$ref: '#/components/responses/NotFound'}
        '413':
          description: 文件超过允许大小
          content:
            application/json:
              schema: {$ref: ./schemas.yaml#/components/schemas/ErrorResponse}
        '415':
          description: 不支持的文件类型
          content:
            application/json:
              schema: {$ref: ./schemas.yaml#/components/schemas/ErrorResponse}
        '422': {$ref: '#/components/responses/ValidationFailed'}
  /api/source-versions/{version_id}:
    parameters:
      - $ref: '#/components/parameters/SourceVersionId'
    get:
      operationId: getSourceVersion
      tags: [Sources]
      summary: 读取资料版本及处理状态
      x-tickets: ['03', '05']
      responses:
        '200':
          description: 资料版本
          content:
            application/json:
              schema: {$ref: ./schemas.yaml#/components/schemas/SourceVersion}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/source-versions/{version_id}/anchors/{anchor_id}:
    parameters:
      - $ref: '#/components/parameters/SourceVersionId'
      - $ref: '#/components/parameters/AnchorId'
    get:
      operationId: getSourceAnchor
      tags: [Sources]
      summary: 读取稳定来源位置及上下文
      x-tickets: ['04', '05']
      responses:
        '200':
          description: 来源位置
          content:
            application/json:
              schema: {$ref: ./schemas.yaml#/components/schemas/SourceAnchor}
        '404': {$ref: '#/components/responses/NotFound'}
        '410':
          description: 资料不可用，但历史引用仍存在
          content:
            application/json:
              schema: {$ref: ./schemas.yaml#/components/schemas/ErrorResponse}
  /api/source-versions/{version_id}/assets/{asset_id}:
    parameters:
      - $ref: '#/components/parameters/SourceVersionId'
      - $ref: '#/components/parameters/AssetId'
    get:
      operationId: getSourceAsset
      tags: [Sources]
      summary: 读取资料版本中的图片或二进制资源
      x-tickets: ['05', '12']
      responses:
        '200':
          description: 原始资源字节
          content:
            image/*:
              schema: {type: string, format: binary}
            application/octet-stream:
              schema: {type: string, format: binary}
        '404': {$ref: '#/components/responses/NotFound'}
        '410':
          description: 资料不可用
          content:
            application/json:
              schema: {$ref: ./schemas.yaml#/components/schemas/ErrorResponse}
  /api/citations/{citation_id}:
    parameters:
      - $ref: '#/components/parameters/CitationId'
    get:
      operationId: getCitation
      tags: [Sources]
      summary: 解析来源引用并返回可复查上下文
      x-tickets: ['04', '05']
      responses:
        '200':
          description: 来源引用；资料不可用时 available=false，身份和位置仍保留
          content:
            application/json:
              schema: {$ref: ./schemas.yaml#/components/schemas/Citation}
        '404': {$ref: '#/components/responses/NotFound'}
...
  /api/operations/{operation_id}:
    parameters:
      - $ref: '#/components/parameters/OperationId'
    get:
      operationId: getOperation
      tags: [Operations]
      summary: 查询异步任务
      x-tickets: ['02', '03', '05', '07', '08', '09', '11', '13', '14']
      responses:
        '200':
          description: 异步任务
          content:
            application/json:
              schema: {$ref: ./schemas.yaml#/components/schemas/Operation}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/operations/{operation_id}/cancel:
    parameters:
      - $ref: '#/components/parameters/OperationId'
    post:
      operationId: cancelOperation
      tags: [Operations]
      summary: 取消可取消的异步任务
      x-tickets: ['02', '03', '05', '07', '08', '09', '11', '13', '14']
      responses:
        '202': {$ref: '#/components/responses/OperationAccepted'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
...
  /api/source-versions/{version_id}/anchors:
    parameters:
      - $ref: '#/components/parameters/SourceVersionId'
    get:
      operationId: listSourceVersionAnchors
      tags: [Sources]
      summary: 列出资料版本锚点
      description: 前端读取来源位置必须使用此接口，不从资料版本摘要猜测 anchors 字段。
      x-tickets: ['16']
      responses:
        '200':
          description: 资料版本锚点
          content:
            application/json:
              schema: {$ref: ./schemas.yaml#/components/schemas/SourceAnchorList}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '410': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
...
components:
  parameters:
    SubjectId: {name: subject_id, in: path, required: true, schema: {$ref: ./schemas.yaml#/components/schemas/Identifier}}
...
    SourceId: {name: source_id, in: path, required: true, schema: {$ref: ./schemas.yaml#/components/schemas/Identifier}}
    SourceVersionId: {name: version_id, in: path, required: true, schema: {$ref: ./schemas.yaml#/components/schemas/Identifier}}
    AnchorId: {name: anchor_id, in: path, required: true, schema: {$ref: ./schemas.yaml#/components/schemas/Identifier}}
    AssetId: {name: asset_id, in: path, required: true, schema: {$ref: ./schemas.yaml#/components/schemas/Identifier}}
    CitationId: {name: citation_id, in: path, required: true, schema: {$ref: ./schemas.yaml#/components/schemas/Identifier}}
...
    OperationId: {name: operation_id, in: path, required: true, schema: {$ref: ./schemas.yaml#/components/schemas/Identifier}}
...
  responses:
    OperationAccepted:
      description: 已接受异步任务
      content:
        application/json:
          schema: {$ref: ./schemas.yaml#/components/schemas/OperationAccepted}
    NotFound:
      description: 资源不存在或不属于当前科目空间
      content:
        application/json:
          schema: {$ref: ./schemas.yaml#/components/schemas/ErrorResponse}
    Conflict:
      description: 当前资源状态不允许此操作
      content:
        application/json:
          schema: {$ref: ./schemas.yaml#/components/schemas/ErrorResponse}
    ValidationFailed:
      description: 请求字段或领域约束校验失败
      content:
        application/json:
          schema: {$ref: ./schemas.yaml#/components/schemas/ErrorResponse}